requirements.txt
functions-framework==3.*
google-cloud-speech==2.21.0
google-auth==2.15.0
